LPS-104504 Inline NullClusterable to ClusterableProxyFactory.
